Menambahkan expired time pada jwt dan juga menyelesaikan add Product route , menambahkan helper untuk cloudinary

+++ toko_sembako_acen_api/models/products.go
@@ -9,7 +9,7 @@ import (
 type Product struct {
 	Id         uuid.UUID  `json:"id" gorm:"type:uuid;not null ; default:gen_random_uuid()"`
 	Name       string     `json:"name" gorm:"type:varchar(255)";not null;`
-	Stock      int64      `json:"stock" gorm:"type:int4";not null;default:0`
+	Stock      int        `json:"stock" gorm:"type:int4";not null;default:0`
 	Price      float64    `json:"price" gorm:"type:float8";not null;`
 	Capital    float64    `json:"capital" gorm:"type:float8;not null"`
 	Picture    string     `json:"picture" gorm:"type:text;not null"`
